refactor: 데이터 구조 수정
