+++ src/components/FilterGroup.js
@@ -1,17 +1,16 @@
-import { CATEGORY_OPTIONS, SORT_OPTIONS } from '../../public/restaurantData.js';
-import { CATEGORY } from '../constants/SETTING.js';
+import { CATEGORY, ORDER } from '../constants/SETTING.js';
 import { createSelectBox } from './common/SelectBox.js';
 
 function createFilterGroup(onChangeCategory, onChangeFilter) {
   const fragment = new DocumentFragment();
   const categoryFilter = createSelectBox({
-    options: [CATEGORY.ALL, ...CATEGORY_OPTIONS],
+    options: Object.values(CATEGORY),
     type: 'category',
     onChange: onChangeCategory,
   });
 
   const sortFilter = createSelectBox({
-    options: SORT_OPTIONS,
+    options: Object.values(ORDER),
     type: 'sorting',
     onChange: onChangeFilter,
   });
